feat: update `looksLikeFlowFileAnnotation` regex to match only @flow (#233)
The `looksLikeFlowFileAnnotation` regex was too broad and would catch
jsdoc comments like `@fixable` and `@function`, causing eslint errors
in projects that don't use flow (in my case).

I updated the regex to only check for anything beginning with
`@flo`. Is there any reason not to full on match `@flow`? I'm not
familiar with how flow works.

This should fix #165.